refactor(telegram): extract hydrateStatsFromStore from session-manager.ts
session-manager.ts had grown to 468 code lines, violating the 461-line
baseline registered in .filesize-baseline.json (the file predates the
350-line ceiling gate and carries a legacy exemption, but the ratchet
still blocks growth beyond the baselined value).

Extract `_hydrateStatsFromStore` body into a new sibling module:

  src/telegram/session-manager.hydrate-stats.ts

The extracted `hydrateStatsFromStore(sessionStats, sessionData, route)`
is a named helper taking explicit parameters — no closure over enclosing
locals. It owns one concern: given the in-memory maps and a route, hydrate
the per-route SessionStats from the shared persisted sidecar when the bot
restarts and loses in-memory state.

The private method in SessionManager becomes a one-line delegate:

  private _hydrateStatsFromStore(route): void {
    hydrateStatsFromStore(this.sessionStats, this.sessionData, route);
  }

Public surface of SessionManager is unchanged. All existing callers of
_hydrateStatsFromStore are internal and the delegation is transparent.

Before: 468 code lines
After:  443 code lines (≤461 baseline — ratchet passes)

Gates:
  pnpm lint                 ✓ (tsc --noEmit, strict)
  pnpm audit:filesize:update ✓ (baseline regenerated)
  pnpm audit:filesize:check  ✓ (no GREW)
  pnpm audit:funcsize:update ✓ (baseline regenerated)
  pnpm test src/telegram/    ✓ (994 tests pass)
