+++ .funcsize-baseline.json
@@ -29,18 +29,18 @@
     "src/cli/_lib/stream-renderer-orchestrator.ts::handleOrchestratorEvent": { "loc": 323, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/_lib/stream-renderer-process.ts::processEvent": { "loc": 201, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/_lib/stream-renderer-subagent.ts::handleSubagentEvent": { "loc": 267, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/cli/commands/chat.ts::registerChatCommand": { "loc": 702, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/cli/commands/chat.ts::registerChatCommand": { "loc": 708, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/config-command.ts::registerConfigCommand": { "loc": 204, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/cli/commands/daemon.ts::registerDaemonCommand": { "loc": 293, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/cli/commands/daemon.ts::registerDaemonCommand": { "loc": 298, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/farm.ts::runFarm": { "loc": 243, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive.ts::registerInteractiveCommand": { "loc": 743, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/cli/commands/interactive/bootstrap.ts::bootstrapSession": { "loc": 332, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/cli/commands/interactive/bootstrap.ts::bootstrapSession": { "loc": 336, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive/footer-subsystems.ts::setupFooterSubsystems": { "loc": 244, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive/loop-iteration.ts::runInputLoop": { "loc": 728, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive/surface-setup.ts::setupSurface": { "loc": 307, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive/tool-lane-render-children.ts::renderOverlayChildren": { "loc": 248, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive/tool-lane.ts::ToolLane.getOverlay": { "loc": 234, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/cli/commands/interactive/turn-handler.ts::runTurn": { "loc": 865, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/cli/commands/interactive/turn-handler.ts::runTurn": { "loc": 869, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/interactive/worktree.ts::createWorktreeAt": { "loc": 265, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/marketplace.ts::registerMarketplaceCommand": { "loc": 220, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/commands/trace.ts::renderEvent": { "loc": 285, "reason": "legacy: predates the function-size gate; pending helper extraction" },
@@ -54,8 +54,8 @@
     "src/cli/terminal-compositor.input-dispatch.ts::handleEnter": { "loc": 259, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/cli/terminal-compositor.lifecycle.ts::arm": { "loc": 217, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/skills/audit-fit/index.ts::handler": { "loc": 291, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/telegram/bot.ts::TelegramBot.setupHandlers": { "loc": 248, "reason": "legacy: predates the function-size gate; pending helper extraction" },
-    "src/telegram/elicitation-handler.ts::makeTelegramElicitationHandler": { "loc": 403, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/telegram/bot.ts::TelegramBot.setupHandlers": { "loc": 247, "reason": "legacy: predates the function-size gate; pending helper extraction" },
+    "src/telegram/elicitation-handler.ts::makeTelegramElicitationHandler": { "loc": 406, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/telegram/handlers/message.ts::MessageHandler.handlePhoto": { "loc": 230, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/telegram/streaming.ts::streamResponse": { "loc": 286, "reason": "legacy: predates the function-size gate; pending helper extraction" },
     "src/web-server/frontend/view-model.ts::createTranscriptModel": { "loc": 214, "reason": "legacy: predates the function-size gate; pending helper extraction" }
