+++ src/telegram/session-manager.ts
@@ -22,6 +22,7 @@ import { type TelegramRoute, routeKey } from './route.js';
 import { sessionRegistry, type SessionRegistry } from '../agent/session/session-registry.js';
 import { ensureRegistryHandle, archiveRegistryHandle } from './session-manager.registry.js';
 import { resolveActiveRouteForChat } from './session-manager.active-route.js';
+import { hydrateStatsFromStore } from './session-manager.hydrate-stats.js';
 import { promises as fs } from 'fs';
 import { join } from 'path';
 
@@ -404,65 +405,11 @@ export class SessionManager {
    * Hydrate in-memory stats from the persisted sidecar for a chat whose
    * sessionId survived a bot restart in sessionData but whose sessionStats
    * entry was lost (sessionStats is in-memory-only and starts empty on restart).
-   *
-   * Guards:
-   * - No-op when stats already exist in memory (never clobber live state).
-   * - No-op when sessionData carries no sessionId (nothing to hydrate from).
-   * - No-op when the sidecar cannot be loaded (missing / corrupted file).
-   * - No-op when the sidecar is not a telegram sidecar for THIS chat (source
-   *   guard prevents accidentally adopting a CLI sidecar that happens to share
-   *   a sessionId).
-   *
-   * After hydration, setSessionName and recordTelegramTurn both see the full
-   * prior-conversation stats (sessionId, turns, totals), so a rename persists
-   * in-place without forking a duplicate sidecar and turn counts are preserved.
+   * Delegates to the extracted `hydrateStatsFromStore` helper
+   * (session-manager.hydrate-stats.ts).
    */
   private _hydrateStatsFromStore(route: TelegramRoute): void {
-    const key = routeKey(route);
-    // Never clobber live in-memory stats — this is a post-restart-only repair.
-    if (this.sessionStats.has(key)) return;
-
-    const sessionId = this.sessionData.get(key)?.sessionId;
-    if (!sessionId) return;
-
-    const stored = loadSession(sessionId);
-    if (!stored) return;
-
-    // Only hydrate telegram sidecars that belong to THIS chat — prevent
-    // accidentally adopting a CLI sidecar or a different chat's sidecar. The
-    // per-route sidecar is keyed by SDK sessionId, so a topic can only hydrate
-    // its own conversation (each topic's session has a distinct sessionId).
-    if (stored.source !== 'telegram' || stored.telegramChatId !== route.chatId) return;
-
-    // Map StoredSession → SessionStats.
-    // Critical rename: stored.startedAt === SessionStats.sessionStartTime.
-    // Fields not persisted (turnCosts, turnTokens, permissionMode) are
-    // reconstructed as empty/default — they are runtime-only display helpers,
-    // not resumption data. The round-trip contract is: saveSession(hydrated) === the original
-    // sidecar (modulo savedAt timestamp), so a post-hydration persist does NOT
-    // fork a new file.
-    const stats: SessionStats = {
-      sessionId: stored.sessionId,
-      name: stored.name,
-      model: stored.model,
-      source: stored.source,
-      telegramChatId: stored.telegramChatId,
-      sessionStartTime: stored.startedAt,
-      totalTurns: stored.totalTurns,
-      totalCostUsd: stored.totalCostUsd,
-      unpricedTurns: stored.unpricedTurns ?? 0,
-      totalTokens: stored.totalTokens,
-      totalDurationMs: stored.totalDurationMs,
-      turns: stored.turns,
-      // Runtime-only fields — reconstructed as empty defaults.
-      turnCosts: [],
-      turnTokens: [],
-      permissionMode: 'default',
-    };
-    // Carry forward the per-route cwd override if one was set via /cd.
-    const chatCwd = this.sessionData.get(key)?.cwd;
-    if (chatCwd !== undefined) stats.cwd = chatCwd;
-    this.sessionStats.set(key, stats);
+    hydrateStatsFromStore(this.sessionStats, this.sessionData, route);
   }
 
   /**
